+++ src/node/rpc/member_frontend.h
@@ -68,6 +68,15 @@ namespace ccf
   DECLARE_JSON_TYPE(KeyIdInfo)
   DECLARE_JSON_REQUIRED_FIELDS(KeyIdInfo, issuer, cert)
 
+  struct FullMemberDetails : public ccf::MemberDetails
+  {
+    crypto::Pem cert;
+    std::optional<crypto::Pem> public_encryption_key;
+  };
+  DECLARE_JSON_TYPE(FullMemberDetails);
+  DECLARE_JSON_REQUIRED_FIELDS(
+    FullMemberDetails, status, member_data, cert, public_encryption_key);
+
   class MemberEndpoints : public CommonEndpointRegistry
   {
   private:
@@ -486,7 +495,7 @@ namespace ccf
       openapi_info.description =
         "This API is used to submit and query proposals which affect CCF's "
         "public governance tables.";
-      openapi_info.document_version = "2.6.0";
+      openapi_info.document_version = "2.7.0";
     }
 
     static std::optional<MemberId> get_caller_member_id(
@@ -1412,6 +1421,52 @@ namespace ccf
         .install();
 
 #pragma clang diagnostic pop
+
+      using AllMemberDetails = std::map<ccf::MemberId, FullMemberDetails>;
+      auto get_all_members =
+        [this](endpoints::ReadOnlyEndpointContext& ctx, nlohmann::json&&) {
+          auto members = ctx.tx.ro<ccf::MemberInfo>(ccf::Tables::MEMBER_INFO);
+          auto member_certs =
+            ctx.tx.ro<ccf::MemberCerts>(ccf::Tables::MEMBER_CERTS);
+          auto member_public_encryption_keys =
+            ctx.tx.ro<ccf::MemberPublicEncryptionKeys>(
+              ccf::Tables::MEMBER_ENCRYPTION_PUBLIC_KEYS);
+
+          AllMemberDetails response;
+
+          members->foreach(
+            [&response, member_certs, member_public_encryption_keys](
+              const auto& k, const auto& v) {
+              FullMemberDetails md;
+              md.status = v.status;
+              md.member_data = v.member_data;
+
+              const auto cert = member_certs->get(k);
+              if (cert.has_value())
+              {
+                md.cert = cert.value();
+              }
+
+              const auto public_encryption_key =
+                member_public_encryption_keys->get(k);
+              if (public_encryption_key.has_value())
+              {
+                md.public_encryption_key = public_encryption_key.value();
+              }
+
+              response[k] = md;
+              return true;
+            });
+
+          return make_success(response);
+        };
+      make_read_only_endpoint(
+        "/members",
+        HTTP_GET,
+        json_read_only_adapter(get_all_members),
+        ccf::no_auth_required)
+        .set_auto_schema<void, AllMemberDetails>()
+        .install();
     }
   };